requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-functions==1.24.0
azure-functions-runtime==1.0.0
azurefunctions-extensions-base==1.1.0
blockbuster==1.5.24
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
filetype==1.2.0
forbiddenfruit==0.1.4
frozenlist==1.6.0
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-auth==2.39.0
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.0
grpcio-status==1.72.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.61
langchain-google-genai==2.1.3
langchain-openai==0.3.18
langchain-tavily==0.1.6
langchain-text-splitters==0.3.8
langgraph==0.3.34
langgraph-api==0.1.16
langgraph-checkpoint==2.0.24
langgraph-cli==0.2.7
langgraph-prebuilt==0.1.8
langgraph-runtime-inmem==0.0.6
langgraph-sdk==0.1.63
langsmith==0.3.34
MarkupSafe==3.0.3
marshmallow==3.26.1
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
numpy==2.2.5
openai==1.82.0
orjson==3.10.16
ormsgpack==1.9.1
packaging==24.2
pdfminer.six==20251107
pdfplumber==0.11.8
pillow==12.0.0
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
PyJWT==2.10.1
pypdfium2==5.0.0
python-dotenv==1.1.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.1.3
starlette==0.46.2
structlog==25.2.0
tavily-python==0.7.3
tenacity==9.1.2
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.0
zstandard==0.23.0
